ledgerx: convert to TS
